# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: CIDER-EM API
  description: "CIDER-EM Context and Inference baseD ontology alignER. CIDER-EM is\
    \ a word-embedding-based system for monolingual and cross-lingual ontology alignment\
    \ deveoped by the SID group (University of Zaragoza). It evolves the CIDER-CL\
    \ tool, deveoped at OEG (Universidad Politécnica de Madrid) and SID (University\
    \ of Zaragoza) by including the use of word embeddings."
  license:
    name: GNU General Public License version 3
    url: "https://www.gnu.org/licenses/gpl-3.0.html"
  version: 1.0.0-oas3
  x-port: 8080

paths:
  /align/entities:
    post:
      tags:
      - alignment
      description: "Given the URIs of two ontologies and two ontology elements, and\
        \ a language code, return the cosine similarity between the two ontology entities"
      operationId: getEntitySimilarity
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntitySimilarity"
        required: true
      responses:
        "200":
          description: An EntitySimilarity instance
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EntitySimilarity"
  /align/ontologies:
    get:
      tags:
      - alignment
      description: "Given the URIs of two ontologies, return an ontology alignment\
        \ in RDF following the Alignment Format [https://moex.gitlabpages.inria.fr/alignapi/format.html]"
      operationId: ontologyAlignment
      parameters:
      - name: uri1
        in: query
        required: true
        schema:
          type: string
        example: "http://oaei.ontologymatching.org/2011/benchmarks/101/onto.rdf"
      - name: uri2
        in: query
        required: true
        schema:
          type: string
        example: "http://oaei.ontologymatching.org/2011/benchmarks/203/onto.rdf"
      responses:
        "200":
          description: An EntitySimilarity instance
          content:
            '*/*':
              schema:
                type: string
                format: binary
components:
  schemas:
    EntitySimilarity:
      type: object
      properties:
        uri1:
          type: string
        uri2:
          type: string
        onto1:
          type: string
        onto2:
          type: string
        lang:
          type: string
        similarity:
          type: number
          format: double
        sim:
          type: number
          format: double
          writeOnly: true